+++ benches/get_chat_msgs.rs
@@ -5,11 +5,14 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use deltachat::chat::{self, ChatId};
 use deltachat::chatlist::Chatlist;
 use deltachat::context::Context;
+use deltachat::stock_str::StockStrings;
 use deltachat::Events;
 
 async fn get_chat_msgs_benchmark(dbfile: &Path, chats: &[ChatId]) {
     let id = 100;
-    let context = Context::new(dbfile, id, Events::new()).await.unwrap();
+    let context = Context::new(dbfile, id, Events::new(), StockStrings::new())
+        .await
+        .unwrap();
 
     for c in chats.iter().take(10) {
         black_box(chat::get_chat_msgs(&context, *c, 0).await.ok());
@@ -23,7 +26,7 @@ fn criterion_benchmark(c: &mut Criterion) {
         let rt = tokio::runtime::Runtime::new().unwrap();
 
         let chats: Vec<_> = rt.block_on(async {
-            let context = Context::new(Path::new(&path), 100, Events::new())
+            let context = Context::new(Path::new(&path), 100, Events::new(), StockStrings::new())
                 .await
                 .unwrap();
             let chatlist = Chatlist::try_load(&context, 0, None, None).await.unwrap();